+++ example/a.ts
@@ -1,37 +1,16 @@
 import { Elysia, t } from '../src'
-import { retrieveRootParamters, sucrose } from '../src/sucrose'
+import { req } from '../test/utils'
 
-// const parameter = '({ hello: { a }, path })'
-// const result = retrieveRootParamters(parameter)
-
-// console.log(result)
-
-console.log(
-	sucrose({
-		handler: function ({ query }) {
-			query.a
-		},
-		afterHandle: [],
-		beforeHandle: [],
-		error: [
-			function a({
-				query,
-				query: { a, c: d },
-				headers: { hello },
-				...rest
-			}) {
-				query.b
-				rest.query.e
-			},
-			({ query: { f } }) => {}
-		],
-		mapResponse: [],
-		onResponse: [],
-		parse: [],
-		request: [],
-		start: [],
-		stop: [],
-		trace: [],
-		transform: []
+new Elysia()
+  .onError(({ code, error }) => {
+	if(code === "VALIDATION")
+		return code
+  })
+  .get('/query', () => {
+	return 'a'
+  }, {
+	query: t.Object({
+		a: t.String()
 	})
-)
+  })
+  .listen(3000)