+++ src/index.ts
@@ -520,20 +520,20 @@ export default class Elysia<
 							models,
 							additionalProperties: !this.config.normalize,
 							coerce: true,
-							additionalCoerce: stringToStructureCoercions
+							additionalCoerce: stringToStructureCoercions()
 						}),
 						params: getSchemaValidator(cloned.params, {
 							dynamic,
 							models,
 							coerce: true,
-							additionalCoerce: stringToStructureCoercions
+							additionalCoerce: stringToStructureCoercions()
 						}),
 						query: getSchemaValidator(cloned.query, {
 							dynamic,
 							models,
 							normalize,
 							coerce: true,
-							additionalCoerce: stringToStructureCoercions
+							additionalCoerce: stringToStructureCoercions()
 						}),
 						cookie: cookieValidator(),
 						response: getResponseSchemaValidator(cloned.response, {
@@ -565,7 +565,8 @@ export default class Elysia<
 									models,
 									additionalProperties: !normalize,
 									coerce: true,
-									additionalCoerce: stringToStructureCoercions
+									additionalCoerce:
+										stringToStructureCoercions()
 								}
 							))
 						},
@@ -578,7 +579,8 @@ export default class Elysia<
 									dynamic,
 									models,
 									coerce: true,
-									additionalCoerce: stringToStructureCoercions
+									additionalCoerce:
+										stringToStructureCoercions()
 								}
 							))
 						},
@@ -591,7 +593,8 @@ export default class Elysia<
 									dynamic,
 									models,
 									coerce: true,
-									additionalCoerce: stringToStructureCoercions
+									additionalCoerce:
+										stringToStructureCoercions()
 								}
 							))
 						},
@@ -2090,12 +2093,7 @@ export default class Elysia<
 						MergeSchema<Ephemeral['schema'], Metadata['schema']>
 					>
 				>,
-				{
-					decorator: Singleton['decorator']
-					store: Singleton['store']
-					derive: Singleton['derive']
-					resolve: Singleton['resolve']
-				},
+				Singleton,
 				Ephemeral,
 				Volatile
 			>
@@ -2116,8 +2114,11 @@ export default class Elysia<
 	 *     })
 	 * ```
 	 */
-	onError<const Schema extends RouteSchema>(
-		options: { as?: LifeCycleType },
+	onError<
+		const Schema extends RouteSchema,
+		const Scope extends LifeCycleType
+	>(
+		options: { as?: Scope },
 		handler: MaybeArray<
 			ErrorHandler<
 				Definitions['error'],
@@ -2128,14 +2129,43 @@ export default class Elysia<
 						MergeSchema<Ephemeral['schema'], Metadata['schema']>
 					>
 				>,
-				{
-					decorator: Singleton['decorator']
-					store: Singleton['store']
-					derive: {}
-					resolve: {}
-				},
-				Ephemeral,
-				Volatile
+				Scope extends 'global'
+					? {
+							store: Singleton['store']
+							decorator: Singleton['decorator']
+							derive: Singleton['derive'] &
+								Ephemeral['derive'] &
+								Volatile['derive']
+							resolve: Singleton['resolve'] &
+								Ephemeral['resolve'] &
+								Volatile['resolve']
+						}
+					: Scope extends 'scoped'
+						? {
+								store: Singleton['store']
+								decorator: Singleton['decorator']
+								derive: Singleton['derive'] &
+									Ephemeral['derive']
+								resolve: Singleton['resolve'] &
+									Ephemeral['resolve']
+							}
+						: Singleton,
+				Scoped extends 'global'
+					? Ephemeral
+					: {
+							derive: Partial<Ephemeral['derive']>
+							resolve: Partial<Ephemeral['resolve']>
+							schema: Ephemeral['schema']
+						},
+				Scoped extends 'global'
+					? Ephemeral
+					: Scoped extends 'scoped'
+						? Ephemeral
+						: {
+								derive: Partial<Ephemeral['derive']>
+								resolve: Partial<Ephemeral['resolve']>
+								schema: Ephemeral['schema']
+							}
 			>
 		>
 	): this
@@ -5793,11 +5823,6 @@ export default class Elysia<
 	 * Beside benchmark purpose, please use 'handle' instead.
 	 */
 	fetch = (request: Request): MaybePromise<Response> => {
-		if (process.env.NODE_ENV === 'production' && this.config.aot !== false)
-			console.warn(
-				"Performance degradation found. Please call Elysia.compile() before using 'fetch'"
-			)
-
 		return (this.fetch = this.config.aot
 			? composeGeneralHandler(this)
 			: createDynamicHandler(this))(request)