+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 1.1.4 - 23 Jul 2024
+Bug fix:
+- removeColonAlias accidentally slice -2 end index for last parameter
+
+# 1.1.3 - 17 Jul 2024
+Change:
+- sucrose: exact inference name
+- use `mapResponse` instead of `mapCompactResponse` for stream
+	- [#727](https://github.com/elysiajs/elysia/issues/727)
+- defers first stream execution before returning response
+	- [#729](https://github.com/elysiajs/elysia/issues/729)
+- [#722](https://github.com/elysiajs/elysia/issues/722) derive context is not passed to onError
+
+Bug fix:
+- `onError` with scope not being able to infer context type
+
 # 1.1.2 - 16 Jul 2024
 Bug fix:
 - [#724](https://github.com/elysiajs/elysia/issues/724), [bun#12594](https://github.com/oven-sh/bun/issues/12594) sucrose: possibly fix `bun build --compile` not being able to infer first, and last context parameter